feat: Add Windows support for nvme-cli (#29)
* docs: Add comprehensive TODO.md for Windows port status

- Document all excluded plugins with priority ratings
- Detail critical missing device I/O functionality
- List implementable stubs (global context, dry-run, host NQN)
- Clarify fabrics string converters not needed (require Linux kernel)
- Add implementation phases and effort estimates
- Minor cleanup: nvme-rpmb.c whitespace

* feat: Add Windows support for nvme-cli

Enable compilation and linking of nvme-cli on Windows using MinGW-W64.
The build succeeds and produces a functional nvme.exe with help system
and 19 vendor plugins. Device I/O operations are stubbed (see TODO.md).

Core Changes:
- Expanded Windows POSIX compatibility layer in platform/windows.h/.c
  * dirent emulation (opendir/readdir/closedir)
  * POSIX functions: getline, strsep, reallocarray, gmtime_r
  * Signal handling: sigaction, sigemptyset
  * Memory mapping stubs: mmap, munmap, madvise
  * DLL loading: dlopen/dlsym/dlclose wrappers
  * mkdir wrapper, readlink stub, open_memstream stub
  * Fabrics/RPMB command stubs (commands registered but unsupported)

- Add platform/types.h stub that defines Windows equivalents to linux/types.h.

- Extend libnvme/src/nvme/windows-stubs.c with 120+ stub functions
  * Keyring/TLS functions (nvme_read_key, nvme_insert_tls_key_*, etc.)
  * Fabrics string converters (nvmf_trtype_str, nvmf_sectype_str, etc.)
  * Tree/topology functions (nvme_scan_topology, nvme_create_root, etc.)
  * Device operations (nvme_open, nvme_close)
  * Host NQN generation (nvmf_hostnqn_generate*)
  * NBFT, global context, path properties

Build System Changes:
- Update meson.build to exclude fabrics.c and nvme-rpmb.c on Windows
- Add windows-stubs.c to Windows-specific sources
- Add Windows library dependencies (ws2_32, wsock32, kernel32, iphlpapi)
- Conditional link args (-ldl on Linux only)

- Update plugins/meson.build:
  * 19 cross-platform plugins enabled (including micron)
  * 10 Linux-only plugins excluded (huawei, ibm, netapp, sandisk,
    scaleflux, wdc, ymtc, zns, lm)

Code Modifications:
- Changed #include of linux/includes.h to platform/includes.h accross the project.
- Add min/max macro guards in common.h to prevent Windows conflicts

Known Limitations:
- All device I/O operations are stubbed (ioctl returns -ENOSYS)
- Cannot open NVMe devices or enumerate devices
- fabrics and rpmb commands print 'not supported on Windows'
- See TODO.md for detailed implementation roadmap

* Fix Windows compilation warnings

- plugins/intel/intel-nvme.c: Use uintptr_t for pointer-to-int casts for 64-bit compatibility
- plugins/micron/micron-nvme.c: Guard min macro definition to avoid redefinition warnings
- plugins/micron/micron-nvme.c: Replace %T with %H:%M:%S in strftime for Windows compatibility
- nvme-print-stdout.c: Fix array-bounds warnings with proper union-based type conversion

* Consolidate Windows platform headers

Reorganizes Windows compatibility layer for cleaner separation between
OS-level abstractions (libnvme) and application-specific code (nvme-cli).

OS-level abstractions (libnvme platform layer):
- Created libnvme/src/nvme/platform/platform_windows.h
  * Windows-specific includes (winsock2.h, windows.h, dirent.h)
  * POSIX compatibility: getline, strsep, reallocarray, readlink,
    open_memstream, fnmatch
  * Signal handling: sigaction, sigemptyset
  * Error codes: ENAVAIL, EREMOTEIO, EDQUOT, ENOTBLK, etc.
  * ioctl constants: BLKBSZSET, BLKRRPART
  * Mode_t typedef and mkdir wrapper macro

- Created libnvme/src/nvme/platform/platform_windows.c
  * POSIX function implementations for Windows compatibility
  * Added exports to libnvme.map for DLL symbol visibility

Application-level stubs (nvme-cli layer):

- Created windows-stubs.c
  * Application-specific stubs: fabrics_discovery, fabrics_connect,
    fabrics_disconnect, fabrics_config, fabrics_dim, rpmb_cmd_option

Combined platform-specific support for both nvme-cli and libnvme and restructured and renamed some of the platform-specific files.

- Moved nvme/platform/*.* into platform/*.* and removed platform_ from file names.
- Moved Windows-specific linux/types.h into platform/types.h and fixed it to support linux.  Updated references to linux/types.h to use platform/types.h.

* Fixed warnings on Windows.

* PATH_MAX is used for some buffers in micron-nvme.c where it is assumed to be larger than 1024.  On Windows it is 260.  On Linux it is typically 4096.  Using 4096.

- Fixed possible underflow error during size comparison.

---------

+++ plugins/intel/intel-nvme.c
@@ -10,7 +10,7 @@
 #include "nvme.h"
 #include "libnvme.h"
 #include "plugin.h"
-#include "linux/types.h"
+#include "platform/types.h"
 #include "nvme-print.h"
 
 #define CREATE_CMD
@@ -1271,7 +1271,7 @@ static int read_header(struct nvme_passthru_cmd *cmd, __u8 *buf,
 	cmd->cdw10 = 0x400;
 	cmd->cdw12 = dw12;
 	cmd->data_len = 0x1000;
-	cmd->addr = (unsigned long)(void *)buf;
+	cmd->addr = (uintptr_t)(void *)buf;
 	return read_entire_cmd(cmd, 0x400, 0x400, -1, hdl, buf);
 }
 
@@ -1416,7 +1416,7 @@ static int get_internal_log(int argc, char **argv, struct command *acmd,
 	/* for 1.1 Fultondales will use old nlog, but current assert/event */
 	if ((intel->ver.major < 1 && intel->ver.minor < 1) ||
 	    (intel->ver.major <= 1 && intel->ver.minor <= 1 && cfg.log == 0)) {
-		cmd.addr = (unsigned long)(void *)buf;
+		cmd.addr = (uintptr_t)(void *)buf;
 		err = get_internal_log_old(buf, output, hdl, &cmd);
 		goto out;
 	}
